backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
openpyxl==3.1.5
python-dotenv==1.0.1
python-pptx==0.6.23
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
